refactor: split profile and public bundle repositories
